Input processing and pipline modifications to allow combined DNA to Amnio comparison
New File fsphmmert.c is a modified version of nhmmer.c. Changes have
been made to the input parsing and processing to ensure correct alphabets
on the target and query (DNA and Aminos respectivly). This stil needs to
be tested to see if it was sucessful.

The only other modification is to comment out the thread_loop option so
that flow always goes to serial_loop(). This way I can focus on this
simplee approach and then come back to the thread version.

p7_pipeline.c has been modified to add a new pipeline function,
P7_Pipeline_Frameshift(). This was copied from P7_Pipeline() and only
the header comments have been modified so far. It will take the same
agruments but the target sequence will be DNA. The pipeline will then
preform the translation functions previously done in phmmert.c and then
run the P7_Pipeline() versions of MSV, bias, and viterbi filters.  Any
amino acid sequences that make it through these filters will have to be
back translated to the original sequence so that a window can be
extracted and passed on the forward as is done in
P7_Pipeline_LongTarget(). The versions of forward, backward, etc. that
will be used at this point will be the frameshift versions that have get
to be writen.