small changes to fsphmmert up to pipeline call
I have tested the alphabet checks on the hmm query input.  I still need
to test the check on MSA or sequence queries. There does not seem to be
a alphabet check on the target in either nhmmer or phmmert so I will not
include on in fsphmmert.

I have followed and modified the flow of control in fsphmmert up to
where the new pipeline is called as p7_Pipeline_Frameshift(). I have
also changed this pipeline to be a copy of p7_Pipeline_LongTarget()
rather that p7_Pipeline(). This is because I will need the extra
parameters passed to the long target version. The next step will be to
begin to modify p7_Pipeline_Frameshift() to combine the long target and
standard filtering steps.