Changes to fsphmmert to include transaltion.
I have decided that it is better to do the translation steps in the
serial_master() function of fsphmmert and have the serial_loop()
function be called once per target sequence.  The loop, then, is for the
ORFs, which are passed one at a time to p7_Pipeline_Frameshift(). This
may not end up being the most optimal way to do this but for now it will
suffice. The pipleine function is set to take the parameters it needs
but has an empty body. Next thing to do will be to bring in the filter
steps from the standard pipeline.