Completed frameshift forward algorithim
The new frameshift forward algorithim is complete and functional. The
only problem at the monent has to do with the window of sequence sent to
forward. I have tried to find a better way to find a widow but I think I
just need to use the already available functioality of nhmmer and the
LongTarget pipeline.