Modified pre-forward pipline to supply correct window to foward
I have made sinificant changes to the pipeline, drawing heavily on the
nhmmer LongTarget pipline. The pipeline now runs as follows:
1) The entire original dna sequence (or it's reverse complement) is sent
to p7_Pipeline_Frameshift(). There the sequence is translated into orfs
which are feed one by one through the MSV filters.
2) If the orf passes MSV it is sent to p7_pli_postMSV_Frameshift().
There it is passed through the bias filter and the vitirbi filter at
p7_ViterbiFilter_longtarget(). This returns a windowlist containing the
coodinates of the viterbi hits, if any.
3)If a window passes viterbi it is sent to
p7_pli_ExtendAndBackTranslateWindows(). This functions extends the
windows bansed on it's coodinated and the models max_length. The new
extended window coodinates are then maped back to the coresponding
coodinates on the original dna sequence.
4)The new dna sequence windows are then sent to p7_pli_MergeWindows()
where overlapping windows are merged. These final windows are then sent,
one by one, to p7_pli_postViterbi_Frameshift(). There the correct
subsequence will be extracted from the original dna and sent to the new
frameshift forward algorithim.