Completed Frameshift version of Backward
Both the new forward and the new backward algorithim have been coded up
sucessfully and rigerously debuged.  They are procuding scores that are
very close to one another (within a few tenths of a bit) but not identical.
One possibility is that the five 0 rows may be adding a small amount to
the forward score.  I am going to leave it for now but I will go back
and try using just one 0 row at some point to see if it gets the scores
closer.