First stab at Decoding and OptAcc - needs debugging
I have coded up the posterior and optimal calcuations as well as at
least some of the traceback. However the trace is producing highly
rrronious alignmnets.  It is going into the core model only once to get
fom n to c. I had the same problem when coding up the protoype but came
up with soultions which should be working in the HMMER implementation so
I don't think it has the same cause.  Most likely just bugs rather than
a fundemental algorithmic insufficeny.  I found out how to get hmmer to
process my "tiny" toy exampe for debugging.  By using the --max flag to
get rid of filters and -l 8 to set the min ORF length to let trough the
highest scoring alignment. Next I will write up a print dumper function
for the FS version of the DP matrix. This should give me a clue as to
wether the extra 0 rows are causing troubles.