Changing order of porterior matirx fill
I had changed the direction of the fill of the posterior probabilities
matrix from top to bottom to bottom to top so that I could reuse the
forward martix instad of the backward matrix. It is possible that this
is causing the issues I have been experiencing with strang alignments. I
am going to try creating a third matrix so that I can go top to bottom
and see if that helps.