Fixed the alignment going in and out of model
I found several small bugs and made two larger changes.  The first was
to create a third matrix so that I could fill the posterior matrix top
to bottom.  I am not sure that this actually made much differnce and
will go back to the old form (filled bottom to top so I can reuse the
fwd matrix) and see if it breaks anything. The major change was that I
added the p7_ReconfigUnihit() function before running the opt_acc
functions.  This prevents the J state from coming from the E state and
makes the J state unusable. This is only a temporary solution. I now see
the reasoning befing the domain def functions. This will run fwd/back on
each J state seperated domin in unihit mode and then send them to
posteior and opt_acc. I will need to implement this for Frameshift.