Pipeline wokring up to alignment output
I have fixed several problems with the domain deffinition aspects of the
pipeline and managed to get the program to display output.  However the
display is full of errors and sometimes causes a seg fault. I am now
looking into the alignment output to trace back the data and see where
the mistakes are made.