Good alignments from changes to match state traceback
I hunted down two sources of the strang alignments I was getting.  The
first was the function reparameterize_model_frameshift.  This was
midified from an earlyier function without much thought.  I'm not sure
that I need it at all so for now it is just commented out.  I will
discuss with Travis if it should be fixed and reintegrated.

The other problem was with the traceback select_m function.  I was
relying entirly on the max function to find both the state and the
codon, but when I looked back at my perl code I saw that I had first
used the posterior matrix to determin the length of the codon and
then only used max to pick the state.

I am now printing correct alignments, but I am also getting a now and
then seg fault that I will have to hunt down.