Fixed Memory Leaks with Valgrind
I used valgrind to debug memory leaks and other probelms.  I am still
getting two errors.  One is an invalid read in Codon_Emissions_Create().
It goes away when I allocate M+2 instead of M+1, which doesn't qhuite
make sense. The other is a conditional jump error.  Travis and I
investigated this error and could not find it's source.  He informed me
that valgrind is notoriously buggy on macs. The next step is to upload
the program to the surver and try running valgrind there to see if I get
the same problems.