Elminated false domains
I have had to make significant changes to the domain decoding and definitoin routines to get them to produce the correct domains.

For the decoding I have had to look into the core model to determine its share of the probability rather than just preforming a 1 - njcp. This is because the probability of i being in the core modle is not contained only in the ith row and the probailitie of a row do not sum to 1.  by getting the true sum of the row I can then determine the proportion constained in the core model.
This creates a big difference in the mocc matrix between on frame and off frame indices.

For the deffinition I then look to all three frames and only call domains that are above threshhold in all 3. This prevents poor domains like the second one from tufA.  Same goes for calling the begining and ending of domains- all three frames must agree.  This prevents micro-domains from being called.