Eliminated seg fault
I tracked down and resolved the seg fault found when using a chromosome length target sequence.  It was caused by the setting of the domain array in the P7_DOMAINDEF struct to NULL. By simply useing the domaindef_reuse function after setting it to NULL the seg fault was eliminated.