Fixes to both fathmm and hmmsearcht
The way fathmm was deciding between the frameshift pipeline and the orf pipelinerequired that the window passed the forward parser which caused ORFs to be trown away unnessicarialy. Now any ORF that cna pass the forward parser on its own AND has a lower p-value than full window gets passed on. For ORFs that do get passed on each domainis now treated as it's own hit. This resloves the isseus with haveing hits reproted in different ways fromthe different pipelines. Finally the hmmsearcht tabluar output was not checking id domains had passed the thresholds before repoting them. I added this check and reported the problem to Travis so he can address in on the develop branch.